+++ SleepModel/HEMSensorChartContainer.h
@@ -9,6 +9,17 @@
 #import <Charts/Charts-Swift.h>
 #import <UIKit/UIKit.h>
 
+@class HEMSensorChartContainer;
+
+@protocol HEMSensorChartScrubberDelegate <NSObject>
+
+- (void)willBeginScrubbingIn:(HEMSensorChartContainer*)chartContainer;
+- (void)didEndScrubbingIn:(HEMSensorChartContainer*)chartContainer;
+- (void)didMoveScrubberTo:(CGPoint)pointInChartView
+                   within:(HEMSensorChartContainer*)chartContainer;
+
+@end
+
 @interface HEMSensorChartContainer : UIView
 
 @property (weak, nonatomic) IBOutlet UILabel* topLimitLabel;
@@ -17,7 +28,10 @@
 @property (weak, nonatomic) IBOutlet UILabel* botLimitLabel;
 @property (weak, nonatomic) IBOutlet UIView* botLimitLine;
 
+@property (nonatomic, weak) id<HEMSensorChartScrubberDelegate> delegate;
+
 - (void)setChartView:(ChartViewBase*)chartView;
 - (ChartViewBase*)chartView;
+- (void)setScrubberColor:(UIColor*)color;
 
 @end